+++ weixin-java-common/src/main/java/me/chanjar/weixin/common/util/http/SimplePostRequestExecutor.java
@@ -27,14 +27,14 @@ public class SimplePostRequestExecutor implements RequestExecutor<String, String
 
   @Override
   public String execute(RequestHttp requestHttp, String uri, String postEntity) throws WxErrorException, IOException {
-    if (requestHttp.getHttpClient() instanceof CloseableHttpClient) {
-      CloseableHttpClient httpClient = (CloseableHttpClient) requestHttp.getHttpClient();
-      HttpHost httpProxy = (HttpHost) requestHttp.getHttpProxy();
+    if (requestHttp.getRequestHttpClient() instanceof CloseableHttpClient) {
+      CloseableHttpClient httpClient = (CloseableHttpClient) requestHttp.getRequestHttpClient();
+      HttpHost httpProxy = (HttpHost) requestHttp.getRequestHttpProxy();
       return executeApache(httpClient, httpProxy, uri, postEntity);
     }
-    if (requestHttp.getHttpClient() instanceof HttpConnectionProvider) {
-      HttpConnectionProvider provider = (HttpConnectionProvider) requestHttp.getHttpClient();
-      ProxyInfo proxyInfo = (ProxyInfo) requestHttp.getHttpProxy();
+    if (requestHttp.getRequestHttpClient() instanceof HttpConnectionProvider) {
+      HttpConnectionProvider provider = (HttpConnectionProvider) requestHttp.getRequestHttpClient();
+      ProxyInfo proxyInfo = (ProxyInfo) requestHttp.getRequestHttpProxy();
       return executeJodd(provider, proxyInfo, uri, postEntity);
     } else {
       //这里需要抛出异常，需要优化
