+++ weixin-java-mp/src/main/java/me/chanjar/weixin/mp/api/impl/apache/WxMpServiceImpl.java
@@ -12,12 +12,9 @@ import me.chanjar.weixin.common.session.StandardSessionManager;
 import me.chanjar.weixin.common.session.WxSessionManager;
 import me.chanjar.weixin.common.util.RandomUtils;
 import me.chanjar.weixin.common.util.crypto.SHA1;
-import me.chanjar.weixin.common.util.http.RequestExecutor;
-import me.chanjar.weixin.common.util.http.URIUtil;
+import me.chanjar.weixin.common.util.http.*;
 import me.chanjar.weixin.common.util.http.apache.ApacheHttpClientBuilder;
 import me.chanjar.weixin.common.util.http.apache.DefaultApacheHttpClientBuilder;
-import me.chanjar.weixin.common.util.http.apache.SimpleGetRequestExecutor;
-import me.chanjar.weixin.common.util.http.apache.SimplePostRequestExecutor;
 import me.chanjar.weixin.mp.api.*;
 import me.chanjar.weixin.mp.api.impl.*;
 import me.chanjar.weixin.mp.bean.*;
@@ -34,7 +31,7 @@ import org.slf4j.LoggerFactory;
 import java.io.IOException;
 import java.util.concurrent.locks.Lock;
 
-public class WxMpServiceImpl implements WxMpService<CloseableHttpClient, HttpHost> {
+public class WxMpServiceImpl implements WxMpService,RequestHttp {
 
   private static final JsonParser JSON_PARSER = new JsonParser();
 
@@ -248,8 +245,8 @@ public class WxMpServiceImpl implements WxMpService<CloseableHttpClient, HttpHos
 
   private WxMpOAuth2AccessToken getOAuth2AccessToken(StringBuilder url) throws WxErrorException {
     try {
-      RequestExecutor<String, CloseableHttpClient, HttpHost, String> executor = new SimpleGetRequestExecutor();
-      String responseText = executor.execute(this.getHttpclient(), this.httpProxy, url.toString(), null);
+      RequestExecutor<String, String> executor = new SimpleGetRequestExecutor();
+      String responseText = executor.execute(this, url.toString(), null);
       return WxMpOAuth2AccessToken.fromJson(responseText);
     } catch (IOException e) {
       throw new RuntimeException(e);
@@ -292,8 +289,8 @@ public class WxMpServiceImpl implements WxMpService<CloseableHttpClient, HttpHos
     }
 
     try {
-      RequestExecutor<String, CloseableHttpClient, HttpHost, String> executor = new SimpleGetRequestExecutor();
-      String responseText = executor.execute(getHttpclient(), this.httpProxy, url.toString(), null);
+      RequestExecutor<String, String> executor = new SimpleGetRequestExecutor();
+      String responseText = executor.execute(this, url.toString(), null);
       return WxMpUser.fromJson(responseText);
     } catch (IOException e) {
       throw new RuntimeException(e);
@@ -308,8 +305,8 @@ public class WxMpServiceImpl implements WxMpService<CloseableHttpClient, HttpHos
     url.append("&openid=").append(oAuth2AccessToken.getOpenId());
 
     try {
-      RequestExecutor<String, CloseableHttpClient, HttpHost, String> executor = new SimpleGetRequestExecutor();
-      executor.execute(getHttpclient(), this.httpProxy, url.toString(), null);
+      RequestExecutor<String, String> executor = new SimpleGetRequestExecutor();
+      executor.execute(this, url.toString(), null);
     } catch (IOException e) {
       throw new RuntimeException(e);
     } catch (WxErrorException e) {
@@ -345,7 +342,7 @@ public class WxMpServiceImpl implements WxMpService<CloseableHttpClient, HttpHos
    * 向微信端发送请求，在这里执行的策略是当发生access_token过期时才去刷新，然后重新执行请求，而不是全局定时请求
    */
   @Override
-  public <T, E> T execute(RequestExecutor<T, CloseableHttpClient, HttpHost, E> executor, String uri, E data) throws WxErrorException {
+  public <T, E> T execute(RequestExecutor<T, E> executor, String uri, E data) throws WxErrorException {
     int retryTimes = 0;
     do {
       try {
@@ -379,8 +376,7 @@ public class WxMpServiceImpl implements WxMpService<CloseableHttpClient, HttpHos
     throw new RuntimeException("微信服务端异常，超出重试次数");
   }
 
-  @Override
-  public synchronized <T, E> T executeInternal(RequestExecutor<T, CloseableHttpClient, HttpHost, E> executor, String uri, E data) throws WxErrorException {
+  public synchronized <T, E> T executeInternal(RequestExecutor<T, E> executor, String uri, E data) throws WxErrorException {
     if (uri.indexOf("access_token=") != -1) {
       throw new IllegalArgumentException("uri参数中不允许有access_token: " + uri);
     }
@@ -390,7 +386,7 @@ public class WxMpServiceImpl implements WxMpService<CloseableHttpClient, HttpHos
     uriWithAccessToken += uri.indexOf('?') == -1 ? "?access_token=" + accessToken : "&access_token=" + accessToken;
 
     try {
-      return executor.execute(getHttpclient(), getHttpProxy(), uriWithAccessToken, data);
+      return executor.execute(this, uriWithAccessToken, data);
     } catch (WxErrorException e) {
       WxError error = e.getError();
       /*
@@ -417,12 +413,12 @@ public class WxMpServiceImpl implements WxMpService<CloseableHttpClient, HttpHos
     }
   }
 
-  @Override
+  //@Override
   public HttpHost getHttpProxy() {
     return this.httpProxy;
   }
 
-  @Override
+  //@Override
   public CloseableHttpClient getHttpclient() {
     return this.httpClient;
   }
@@ -525,5 +521,15 @@ public class WxMpServiceImpl implements WxMpService<CloseableHttpClient, HttpHos
   @Override
   public WxMpDeviceService getDeviceService() {
     return this.deviceService;
+  }
+
+  @Override
+  public Object getRequestHttpClient() {
+    return this.httpClient;
+  }
+
+  @Override
+  public Object getRequestHttpProxy() {
+    return this.httpProxy;
   }
 }
