--- weixin-java-mp/src/main/java/me/chanjar/weixin/mp/api/impl/jodd/WxMpUserBlacklistServiceImpl.java
@@ -1,52 +1,0 @@
-package me.chanjar.weixin.mp.api.impl.jodd;
-
-import com.google.gson.Gson;
-import com.google.gson.JsonObject;
-import jodd.http.HttpConnectionProvider;
-import jodd.http.ProxyInfo;
-import me.chanjar.weixin.common.exception.WxErrorException;
-import me.chanjar.weixin.common.util.http.jodd.SimplePostRequestExecutor;
-import me.chanjar.weixin.mp.api.WxMpService;
-import me.chanjar.weixin.mp.api.WxMpUserBlacklistService;
-import me.chanjar.weixin.mp.bean.result.WxMpUserBlacklistGetResult;
-
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-/**
- * @author miller
- */
-public class WxMpUserBlacklistServiceImpl implements WxMpUserBlacklistService {
-  private static final String API_BLACKLIST_PREFIX = "https://api.weixin.qq.com/cgi-bin/tags/members";
-  private WxMpService<HttpConnectionProvider, ProxyInfo> wxMpService;
-
-  public WxMpUserBlacklistServiceImpl(WxMpService wxMpService) {
-    this.wxMpService = wxMpService;
-  }
-
-  @Override
-  public WxMpUserBlacklistGetResult getBlacklist(String nextOpenid) throws WxErrorException {
-    JsonObject jsonObject = new JsonObject();
-    jsonObject.addProperty("begin_openid", nextOpenid);
-    String url = API_BLACKLIST_PREFIX + "/getblacklist";
-    String responseContent = this.wxMpService.execute(new SimplePostRequestExecutor(), url, jsonObject.toString());
-    return WxMpUserBlacklistGetResult.fromJson(responseContent);
-  }
-
-  @Override
-  public void pushToBlacklist(List<String> openidList) throws WxErrorException {
-    Map<String, Object> map = new HashMap<>();
-    map.put("openid_list", openidList);
-    String url = API_BLACKLIST_PREFIX + "/batchblacklist";
-    this.wxMpService.execute(new SimplePostRequestExecutor(), url, new Gson().toJson(map));
-  }
-
-  @Override
-  public void pullFromBlacklist(List<String> openidList) throws WxErrorException {
-    Map<String, Object> map = new HashMap<>();
-    map.put("openid_list", openidList);
-    String url = API_BLACKLIST_PREFIX + "/batchunblacklist";
-    this.wxMpService.execute(new SimplePostRequestExecutor(), url, new Gson().toJson(map));
-  }
-}
