--- weixin-java-mp/src/main/java/me/chanjar/weixin/mp/util/http/jodd/MaterialUploadRequestExecutor.java
@@ -1,53 +1,0 @@
-package me.chanjar.weixin.mp.util.http.jodd;
-
-import jodd.http.HttpConnectionProvider;
-import jodd.http.HttpRequest;
-import jodd.http.HttpResponse;
-import jodd.http.ProxyInfo;
-import me.chanjar.weixin.common.bean.result.WxError;
-import me.chanjar.weixin.common.exception.WxErrorException;
-import me.chanjar.weixin.common.util.http.RequestExecutor;
-import me.chanjar.weixin.common.util.json.WxGsonBuilder;
-import me.chanjar.weixin.mp.bean.material.WxMpMaterial;
-import me.chanjar.weixin.mp.bean.material.WxMpMaterialUploadResult;
-
-import java.io.File;
-import java.io.FileNotFoundException;
-import java.io.IOException;
-import java.util.Map;
-
-public class MaterialUploadRequestExecutor implements RequestExecutor<WxMpMaterialUploadResult, HttpConnectionProvider, ProxyInfo, WxMpMaterial> {
-
-  @Override
-  public WxMpMaterialUploadResult execute(HttpConnectionProvider provider, ProxyInfo httpProxy, String uri, WxMpMaterial material) throws WxErrorException, IOException {
-    HttpRequest request = HttpRequest.post(uri);
-    if (httpProxy != null) {
-      provider.useProxy(httpProxy);
-    }
-    request.withConnectionProvider(provider);
-
-    if (material == null) {
-      throw new WxErrorException(WxError.newBuilder().setErrorMsg("非法请求，material参数为空").build());
-    }
-
-    File file = material.getFile();
-    if (file == null || !file.exists()) {
-      throw new FileNotFoundException();
-    }
-    request.form("media", file);
-    Map<String, String> form = material.getForm();
-    if (material.getForm() != null) {
-      request.form("description", WxGsonBuilder.create().toJson(form));
-    }
-
-    HttpResponse response = request.send();
-    String responseContent = response.bodyText();
-    WxError error = WxError.fromJson(responseContent);
-    if (error.getErrorCode() != 0) {
-      throw new WxErrorException(error);
-    } else {
-      return WxMpMaterialUploadResult.fromJson(responseContent);
-    }
-  }
-
-}
