+++ weixin-java-mp/src/test/java/me/chanjar/weixin/mp/api/WxMpBusyRetryTest.java
@@ -4,8 +4,6 @@ import me.chanjar.weixin.common.bean.result.WxError;
 import me.chanjar.weixin.common.exception.WxErrorException;
 import me.chanjar.weixin.common.util.http.RequestExecutor;
 import me.chanjar.weixin.mp.api.impl.apache.WxMpServiceImpl;
-import org.apache.http.HttpHost;
-import org.apache.http.impl.client.CloseableHttpClient;
 import org.testng.annotations.DataProvider;
 import org.testng.annotations.Test;
 
@@ -23,7 +21,7 @@ public class WxMpBusyRetryTest {
 
       @Override
       public synchronized <T, E> T executeInternal(
-        RequestExecutor<T, CloseableHttpClient, HttpHost, E> executor, String uri, E data)
+        RequestExecutor<T, E> executor, String uri, E data)
         throws WxErrorException {
         this.log.info("Executed");
         WxError error = new WxError();
